[reconfigurator] Extract SP related planning code into its own submodule
